Migrate states enum
